Move varmap_to_string and such to Handle.{h,cc}
