+++ opencog/atoms/base/Handle.h
@@ -34,6 +34,7 @@
 #include <set>
 #include <unordered_set>
 #include <vector>
+#include <map>
 
 /** \addtogroup grp_atomspace
  *  @{
@@ -223,6 +224,12 @@ typedef std::set<Handle> OrderedHandleSet;
 //! a hash that associates the handle to its unique identificator
 typedef std::unordered_set<Handle, handle_hash> UnorderedHandleSet;
 
+//! an ordered map from Handle to Handle set
+typedef std::map<Handle, UnorderedHandleSet> VarMultimap;
+
+//! an ordered map from Handle to Handle
+typedef std::map<Handle, Handle> VarMap;
+
 //! a handle iterator
 typedef std::iterator<std::forward_iterator_tag, Handle> HandleIterator;
 
@@ -297,6 +304,8 @@ ostream& operator<<(ostream& out, const opencog::UnorderedHandleSet& hs);
 string hs_to_string(const opencog::HandleSeq& hs);
 string ohs_to_string(const opencog::OrderedHandleSet& ohs);
 string uhs_to_string(const opencog::UnorderedHandleSet& uhs);
+string varmap_to_string(const opencog::VarMap& vm);
+string varmultimap_to_string(const opencog::VarMultimap& vmm);
 
 #ifdef THIS_USED_TO_WORK_GREAT_BUT_IS_BROKEN_IN_GCC472
 // The below used to work, but broke in gcc-4.7.2. The reason it
