+++ opencog/rule-engine/backwardchainer/BackwardChainer.cc
@@ -35,32 +35,6 @@
 
 using namespace opencog;
 
-std::string varmap_to_string(const VarMap& varmap)
-{
-	stringstream ss;
-	int i = 0;
-	for (const auto& p : varmap) {
-		ss << "key[" << i << "]:" << std::endl << p.first->toString()
-		   << "value[" << i << "]:" << std::endl << p.second->toString();
-		i++;
-	}
-	return ss.str();
-}
-
-std::string varmultimap_to_string(const VarMultimap& varmultimap)
-{
-	stringstream ss;
-	int i = 0;
-	for (const auto& p : varmultimap) {
-		ss << "key[" << i << "]:" << std::endl << p.first->toString()
-		   << "value[" << i << "]:" << std::endl;
-		for (const auto s : p.second)
-			ss << s->toString();
-		i++;
-	}
-	return ss.str();
-}
-
 BackwardChainer::BackwardChainer(AtomSpace& as, Handle rbs)
 	: _as(as), _configReader(as, rbs),
 	  // create a garbage superspace with _as as parent, so codes
