+++ opencog/rule-engine/backwardchainer/BackwardChainer.h
@@ -33,13 +33,6 @@ class BackwardChainerUTest;
 
 namespace opencog
 {
-    
-typedef std::map<Handle, UnorderedHandleSet> VarMultimap;
-typedef std::map<Handle, Handle> VarMap;
-
-// Convenient for GDB debugging
-std::string varmap_to_string(const VarMap&);
-std::string varmultimap_to_string(const VarMultimap&);
 
 /**
  * Backward chaining falls into two cases
